fix: store config/pipeline hashes as int64 to fit the uint32 range
The pipeline and Vector config hashes are CRC32 checksums (uint32,
0..4294967295), but their status fields were int32 (*uint32 in Go makes
controller-gen emit format: int32). Any hash above the int32 ceiling
(2147483647) - roughly half of all values - is rejected by the API server:

    status: Invalid value: "": Checked value must be of type integer
    with format int32 in LastAppliedPipelineHash

The status update then fails on every reconcile, so the resource stays
stuck with configCheckResult=false and the reconcile loop retries forever.
VectorPipeline only appeared unaffected when its hash happened to land
below the int32 ceiling; the defect is not specific to ClusterVectorPipeline.

Store the hashes as int64, which fully contains the uint32 range:
  - VectorPipelineStatus.LastAppliedPipelineHash
  - VectorCommonStatus.LastAppliedConfigHash
  - VectorCommonStatus.LastAppliedGlobalConfigHash

int64 is backward compatible with the int32-range values already persisted
in status. CRDs regenerated with controller-gen v0.16.1 and the helm chart
CRD copies synced.

Fixes #232

+++ api/v1alpha1/vectorpipeline_types.go
@@ -36,10 +36,14 @@ type VectorPipelineSpec struct {
 
 // VectorPipelineStatus defines the observed state of VectorPipeline
 type VectorPipelineStatus struct {
-	Role                    *VectorPipelineRole `json:"role,omitempty"`
-	ConfigCheckResult       *bool               `json:"configCheckResult,omitempty"`
-	Reason                  *string             `json:"reason,omitempty"`
-	LastAppliedPipelineHash *uint32             `json:"LastAppliedPipelineHash,omitempty"`
+	Role              *VectorPipelineRole `json:"role,omitempty"`
+	ConfigCheckResult *bool               `json:"configCheckResult,omitempty"`
+	Reason            *string             `json:"reason,omitempty"`
+	// LastAppliedPipelineHash holds the CRC32 (uint32) hash of the last successfully
+	// applied pipeline config. It is stored as an int64 because a uint32 can exceed the
+	// int32 upper bound (2147483647); an int32 field would reject roughly half of all
+	// hash values and leave the pipeline stuck with configCheckResult=false. See #232.
+	LastAppliedPipelineHash *int64 `json:"LastAppliedPipelineHash,omitempty"`
 }
 
 //+kubebuilder:object:root=true
